Resolved linting errors.
