--- src/containers/About/About.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const About = () => <div>About Page</div>
-
-export default About
