--- src/containers/Footer/Footer.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const Footer = () => <div>This is Footer</div>
-
-export default Footer
