--- src/containers/Header/Header.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-import { Logo } from '../../components'
-
-import './Header.css'
-
-const Header = () => (
-  <header className="Header-header">
-    <Logo />
-    <h2>Welcome to KYBING</h2>
-  </header>
-)
-
-export default Header
