+++ src/containers/Header/index.js
@@ -1,2 +1,14 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as Header } from './Header'
+import React from 'react'
+
+import { Logo } from '../../components'
+
+import './styles.css'
+
+const Header = () => (
+  <header className="Header-header">
+    <Logo />
+    <h2>Welcome to KYBING</h2>
+  </header>
+)
+
+export default Header
