--- src/containers/Home/Home.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-import './Home.css'
-
-const Home = () => (
-  <div className="Home-members">
-    <h3>Board of Directors</h3>
-    <ul>Uday Sravan Kumar Kamineni</ul>
-    <ul>Amarendhar Ganji</ul>
-    <ul>Anusha Patil</ul>
-    <ul>Kalyan Pavan Kumar Kamineni</ul>
-    <ul>Saritha Madala</ul>
-  </div>
-)
-
-export default Home
