+++ src/containers/Home/index.js
@@ -1,2 +1,16 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as Home } from './Home'
+import React from 'react'
+
+import './styles.css'
+
+const Home = () => (
+  <div className="Home-members">
+    <h3>Board of Directors</h3>
+    <ul>Uday Sravan Kumar Kamineni</ul>
+    <ul>Amarendhar Ganji</ul>
+    <ul>Anusha Patil</ul>
+    <ul>Kalyan Pavan Kumar Kamineni</ul>
+    <ul>Saritha Madala</ul>
+  </div>
+)
+
+export default Home
