+++ src/containers/index.js
@@ -1,4 +1,6 @@
-export { About } from './About'
-export { Header } from './Header'
-export { Home } from './Home'
-export { Footer } from './Footer'
+import About from './About'
+import Header from './Header'
+import Home from './Home'
+import Footer from './Footer'
+
+export { About, Header, Home, Footer }
